bot/requirements.txt
# Language Model Orchestration
langchain==0.1.5
langchain-openai==0.0.5

# API requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0
